+++ src/main/java/com/telepathicgrunt/blame/main/DynamicRegistriesBlame.java
@@ -92,36 +92,36 @@ public class DynamicRegistriesBlame {
 	public static void printUnregisteredWorldgenConfiguredStuff(net.minecraft.util.registry.DynamicRegistries.Impl imp)
 	{
 		// Create a store here to minimize memory impact and let it get garbaged collected later.
-		Map<String, Set<ResourceLocation>> unconfigured_stuff_map = new HashMap<>();
-		Set<String> collected_possible_issue_mods = new HashSet<>();
+		Map<String, Set<ResourceLocation>> unconfiguredStuffMap = new HashMap<>();
+		Set<String> collectedPossibleIssueMods = new HashSet<>();
 		Gson gson = new GsonBuilder().setPrettyPrinting().create();
 		Pattern pattern = Pattern.compile("\"(?:Name|type|location)\": *\"([a-z_:]+)\"");
 
 		// ConfiguredFeatures
 		imp.func_230521_a_(Registry.field_243552_au).ifPresent(configuredFeatureRegistry ->
 		imp.func_230521_a_(Registry.BIOME_KEY).ifPresent(mutableRegistry -> mutableRegistry.getEntries()
-				.forEach(mapEntry -> findUnregisteredConfiguredFeatures(mapEntry, unconfigured_stuff_map, configuredFeatureRegistry, gson))));
+				.forEach(mapEntry -> findUnregisteredConfiguredFeatures(mapEntry, unconfiguredStuffMap, configuredFeatureRegistry, gson))));
 
-		printUnregisteredStuff(unconfigured_stuff_map, "ConfiguredFeature");
-		extractModNames(unconfigured_stuff_map, collected_possible_issue_mods, pattern);
+		printUnregisteredStuff(unconfiguredStuffMap, "ConfiguredFeature");
+		extractModNames(unconfiguredStuffMap, collectedPossibleIssueMods, pattern);
 
 		// ConfiguredStructures
 		imp.func_230521_a_(Registry.field_243553_av).ifPresent(configuredStructureRegistry ->
 		imp.func_230521_a_(Registry.BIOME_KEY).ifPresent(mutableRegistry -> mutableRegistry.getEntries()
-				.forEach(mapEntry -> findUnregisteredConfiguredStructures(mapEntry, unconfigured_stuff_map, configuredStructureRegistry, gson))));
+				.forEach(mapEntry -> findUnregisteredConfiguredStructures(mapEntry, unconfiguredStuffMap, configuredStructureRegistry, gson))));
 
-		printUnregisteredStuff(unconfigured_stuff_map, "ConfiguredStructure");
-		extractModNames(unconfigured_stuff_map, collected_possible_issue_mods, pattern);
+		printUnregisteredStuff(unconfiguredStuffMap, "ConfiguredStructure");
+		extractModNames(unconfiguredStuffMap, collectedPossibleIssueMods, pattern);
 
 		// ConfiguredCarvers
 		imp.func_230521_a_(Registry.field_243551_at).ifPresent(configuredCarverRegistry ->
 		imp.func_230521_a_(Registry.BIOME_KEY).ifPresent(mutableRegistry -> mutableRegistry.getEntries()
-				.forEach(mapEntry -> findUnregisteredConfiguredCarver(mapEntry, unconfigured_stuff_map, configuredCarverRegistry, gson))));
+				.forEach(mapEntry -> findUnregisteredConfiguredCarver(mapEntry, unconfiguredStuffMap, configuredCarverRegistry, gson))));
 
-		printUnregisteredStuff(unconfigured_stuff_map, "ConfiguredStructure");
-		extractModNames(unconfigured_stuff_map, collected_possible_issue_mods, pattern);
+		printUnregisteredStuff(unconfiguredStuffMap, "ConfiguredStructure");
+		extractModNames(unconfiguredStuffMap, collectedPossibleIssueMods, pattern);
 
-		if(collected_possible_issue_mods.size() != 0){
+		if(collectedPossibleIssueMods.size() != 0){
 			// Add extra info to the log.
 			String errorReport = "\n****************** Blame Report ******************" +
 					"\n\n This is an experimental report. It is suppose to automatically read" +
@@ -129,25 +129,25 @@ public class DynamicRegistriesBlame {
 					"\n and ConfiguredCarvers. Then does its best to collect the terms that seem to" +
 					"\n state whose mod the unregistered stuff belongs to." +
 					"\n\n Possible mods responsible for unregistered stuff: \n" +
-					collected_possible_issue_mods.stream().sorted().collect(Collectors.joining("\n")) + "\n\n";
+					collectedPossibleIssueMods.stream().sorted().collect(Collectors.joining("\n")) + "\n\n";
 
 			// Log it to the latest.log file as well.
 			Blame.LOGGER.log(Level.ERROR, errorReport);
 		}
-		collected_possible_issue_mods.clear();
+		collectedPossibleIssueMods.clear();
 	}
 
-	private static void extractModNames(Map<String, Set<ResourceLocation>> unconfigured_stuff_map, Set<String> collected_possible_issue_mods, Pattern pattern) {
-		unconfigured_stuff_map.keySet()
+	private static void extractModNames(Map<String, Set<ResourceLocation>> unconfiguredStuffMap, Set<String> collectedPossibleIssueMods, Pattern pattern) {
+		unconfiguredStuffMap.keySet()
 				.forEach(jsonString -> {
 					Matcher match = pattern.matcher(jsonString);
 					while(match.find()) {
 						if(!match.group(1).contains("minecraft:")){
-							collected_possible_issue_mods.add(match.group(1));
+							collectedPossibleIssueMods.add(match.group(1));
 						}
 					}
 				});
-		unconfigured_stuff_map.clear();
+		unconfiguredStuffMap.clear();
 	}
 
 	/**
@@ -155,7 +155,7 @@ public class DynamicRegistriesBlame {
 	 */
 	private static void findUnregisteredConfiguredFeatures(
 			Map.Entry<RegistryKey<Biome>, Biome>  mapEntry,
-			Map<String, Set<ResourceLocation>> unregistered_feature_map,
+			Map<String, Set<ResourceLocation>> unregisteredFeatureMap,
 			MutableRegistry<ConfiguredFeature<?,?>> configuredFeatureRegistry,
 			Gson gson)
 	{
@@ -172,7 +172,7 @@ public class DynamicRegistriesBlame {
 							.ifPresent(configuredFeatureJSON ->
 									cacheUnregisteredObject(
 											configuredFeatureJSON,
-											unregistered_feature_map,
+											unregisteredFeatureMap,
 											biomeID,
 											gson));
 				}
@@ -185,7 +185,7 @@ public class DynamicRegistriesBlame {
 	 */
 	private static void findUnregisteredConfiguredStructures(
 		Map.Entry<RegistryKey<Biome>, Biome>  mapEntry,
-		Map<String, Set<ResourceLocation>> unregistered_structure_map,
+		Map<String, Set<ResourceLocation>> unregisteredStructureMap,
 		MutableRegistry<StructureFeature<?,?>> configuredStructureRegistry,
 		Gson gson)
 	{
@@ -200,7 +200,7 @@ public class DynamicRegistriesBlame {
 						.ifPresent(configuredStructureJSON ->
 								cacheUnregisteredObject(
 										configuredStructureJSON,
-										unregistered_structure_map,
+										unregisteredStructureMap,
 										biomeID,
 										gson));
 			}
@@ -212,7 +212,7 @@ public class DynamicRegistriesBlame {
 	 */
 	private static void findUnregisteredConfiguredCarver(
 		Map.Entry<RegistryKey<Biome>, Biome>  mapEntry,
-		Map<String, Set<ResourceLocation>> unregistered_carver_map,
+		Map<String, Set<ResourceLocation>> unregisteredCarverMap,
 		MutableRegistry<ConfiguredCarver<?>> configuredCarverRegistry,
 		Gson gson)
 	{
@@ -228,7 +228,7 @@ public class DynamicRegistriesBlame {
 							.ifPresent(configuredCarverJSON ->
 									cacheUnregisteredObject(
 											configuredCarverJSON,
-											unregistered_carver_map,
+											unregisteredCarverMap,
 											biomeID,
 											gson));
 				}
@@ -238,20 +238,20 @@ public class DynamicRegistriesBlame {
 
 	private static void cacheUnregisteredObject(
 			JsonElement configuredObjectJSON,
-			Map<String, Set<ResourceLocation>> unregistered_object_map,
+			Map<String, Set<ResourceLocation>> unregisteredObjectMap,
 			ResourceLocation biomeID,
 			Gson gson)
 	{
 		String cfstring = gson.toJson(configuredObjectJSON);
 
-		if(!unregistered_object_map.containsKey(cfstring))
-			unregistered_object_map.put(cfstring, new HashSet<>());
+		if(!unregisteredObjectMap.containsKey(cfstring))
+			unregisteredObjectMap.put(cfstring, new HashSet<>());
 
-		unregistered_object_map.get(cfstring).add(biomeID);
+		unregisteredObjectMap.get(cfstring).add(biomeID);
 	}
 
-	private static void printUnregisteredStuff(Map<String, Set<ResourceLocation>> UNREGISTERED_STUFF_MAP, String type){
-		for(Map.Entry<String, Set<ResourceLocation>> entry : UNREGISTERED_STUFF_MAP.entrySet()){
+	private static void printUnregisteredStuff(Map<String, Set<ResourceLocation>> unregisteredStuffMap, String type){
+		for(Map.Entry<String, Set<ResourceLocation>> entry : unregisteredStuffMap.entrySet()){
 
 			// Add extra info to the log.
 			String errorReport = "\n****************** Blame Report ******************" +